Centralize dispatch for determining need for passphrase, errorhandling
Calls to encrypt and decrypt now pass through a single method which will
eventualy handle the variety of cases (are we using symmetric or
asymmetric encryption? do we need a passphrase? should we try again by
prompting for a passphrase? did decryption in fact fail? etc. etc. All
these and more.