the REAL fix to the improper use of __ensure_path
get_path is used ONLY for getting the full path for a cred, and as such
it's the only case which would call __ensure_path with an argument
terminating in a file name and as such it's only proper to call
os.path.dirname(path) on to ensure the existance of that file's parent
path (i.e. a new namespace directory).